added gesture setup to ansible and small edit to scrips

+++ ansible/run.sh
@@ -2,10 +2,14 @@
 
 if [ -e /usr/bin/ansible-playbook ]; then
     sudo ansible-playbook -i hosts applications.yml
-    sudo rm -r $HOME/.ansible # remove ansible config after to prvent perm conflicts
+    if [ -e $HOME/.ansible ]; then 
+        sudo rm -r $HOME/.ansible # remove ansible config after to prvent perm conflicts
+    fi
 
-    ansible-playbook -i hosts configs.yml
-    rm -r $HOME/.ansible # remove ansible dir after just cuz i dont need it
+    ansible-playbook -i hosts configs.yml --tags "link,vim,lock,scripts,rc,cups,terminal,net_timeout,gesture"
+    if [ -e $HOME/.ansible ]; then 
+        rm -r $HOME/.ansible # remove ansible dir after just cuz i dont need it
+    fi
     exit 0
 fi
